A layer of indirection fixes breakpoints
Otherwise the breakpoint is in the wrong place!
