+++ imports/ui/layouts/main/layout.vue
@@ -8,7 +8,6 @@
         <div v-if="logged_in">
           <div v-if="no_harbormasters">
             <new-harbormaster></new-harbormaster>
-            <!-- <div v-blaze="'new_harbormaster'"></div> -->
           </div>
           <div v-else>
             <nav>
@@ -27,7 +26,6 @@
     <div v-else>
       <div v-blaze="'spinner'"></div>
     </div>
-  <!-- <div v-blaze="'main'"></div> -->
   </div>
 </template>
 
@@ -36,7 +34,13 @@ import AddUser from '../../pages/users/add_user';
 import Navigation from '../../components/navigation';
 import NewHarbormaster from '../../components/new_harbormaster';
 import Welcome from '../../components/welcome';
-import { Users } from '../../../api/users';
+
+import {
+  is_loaded,
+  no_users,
+  logged_in,
+  no_harbormasters,
+} from './lib';
 
 const Constraints = new ReactiveVar({});
 
@@ -47,56 +51,10 @@ export default {
       'Lanes': [],
       'Harbors': [],
     },
-    is_loaded () {
-      if (
-        ! Session.get('loading')
-        && ! Meteor.loggingIn()
-      ) {
-        return true;
-      }
-      return false;
-    },
-
-    no_users () {
-      if (! Users.find().fetch().length) { return true; }
-      return false;
-    },
-
-    logged_in () {
-      return Meteor.user();
-    },
-
-    no_harbormasters () {
-      var harbormasters = Users.find({ harbormaster: true }).fetch();
-
-      return ! harbormasters.length ? true : false;
-    },
-
-    // Not yet implemented
-    constraints () {
-      const constraints = Constraints.get();
-      for (let [key, value] of Object.entries(constraints)) {
-        value.forEach((constraint) => {
-          if (constraint.rel) {
-            const link = document.createElement('link');
-            link.rel = constraint.rel;
-            link.href = constraint.href;
-            link.id = constraint.id;
-            document.head.appendChild(link);
-            return link;
-          }
-
-          const script = document.createElement('script');
-          script.async = constraint.async || false;
-          script.id = constraint.id;
-          if (constraint.src) script.src = constraint.src;
-          else if (constraint.text) script.text = constraint.text;
-          else throw new Error('A "src" or "text" field must be supplied!');
-          document.body.appendChild(script);
-          return script;
-        });
-      }
-    },
+    is_loaded,
+    no_users,
+    logged_in,
+    no_harbormasters,
   },
 
   components: {
