+++ imports/ui/pages/lanes/edit_lane/edit_lane.css
@@ -1,0 +1,75 @@
+form p {
+  margin-bottom: 15px;
+}
+
+li {
+  list-style-type: circle;
+  margin: 5px;
+}
+
+.destination[data-destination-index="0"] .remove-destination {
+  display: none;
+}
+
+.address-field {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  display: inline-block;
+}
+
+.destination-address {
+  padding-right: 15px;
+}
+
+.stop {
+  position: relative;
+}
+
+.stop .close-button,
+.address-field .close-button {
+  right: 0;
+  top: 0;
+  background: #eee;
+  width: 40px;
+  height: 39px;
+  border: 1px solid #ccc;
+}
+
+.stop .close-button {
+  top: 25px;
+}
+
+.destination-address[data-address-index="0"] + .remove-address {
+  display: none;
+}
+
+.stop[data-stop-index="0"] .remove-stop {
+  display: none;
+}
+
+fieldset.followup {
+  background: #0af;
+}
+
+fieldset.followup label {
+  color: white;
+}
+
+fieldset.salvage-plan {
+  background: #ffae00;
+}
+
+fieldset.captains ul li {
+  display: inline-block;
+}
+
+.url {
+  margin: 15px 0;
+}
+
+@media all and (max-width: 1200px) {
+  .url {
+    margin: 35px 0;
+  }
+}
